refactor components structure
Moved UI and home feature components to a more organized folder structure under 'common' and 'features'. Centralized shared types (News, Review, Album) into a new 'types.ts' file and updated imports across mock data and components to use these types. This improves maintainability and code organization.
