+++ frontend/src/components/common/SectionTitle.tsx
