+++ frontend/src/components/features/home/news/NewsCard.tsx
