+++ frontend/src/pages/Home.tsx
@@ -1,8 +1,8 @@
 import BaseLayout from "../layouts/BaseLayout";
 import ContentLayout from "../layouts/ContentLayout";
-import TrendingSection from "../components/home/TrendingSection";
-import NewsSection from "../components/home/NewsSection";
-import PopularReviews from "../components/home/PopularReviewsSection";
+import TrendingSection from "../components/features/home/trending-albums/TrendingAlbumsSection";
+import NewsSection from "../components/features/home/news/NewsSection";
+import PopularReviews from "../components/features/home/popular-reviews/PopularReviewsSection";
 
 export default function Home() {
   return (
