+++ frontend/src/pages/Welcome.tsx
@@ -1,9 +1,9 @@
 import BaseLayout from "../layouts/BaseLayout";
 import ContentLayout from "../layouts/ContentLayout";
-import HeroSection from "../components/home/HeroSection";
-import TrendingSection from "../components/home/TrendingSection";
-import NewsSection from "../components/home/NewsSection";
-import PopularReviews from "../components/home/PopularReviewsSection";
+import HeroSection from "../components/features/home/HeroSection";
+import TrendingSection from "../components/features/home/trending-albums/TrendingAlbumsSection";
+import NewsSection from "../components/features/home/news/NewsSection";
+import PopularReviews from "../components/features/home/popular-reviews/PopularReviewsSection";
 
 export default function Welcome() {
   return (
